fix(log_router): do not enter critical section without mutex
Wait as long as it takes to acquire the mutex (previously, it would only
wait 100ms for the mutex before entering the critical section anyway!)

fix(log_router): ensure mutex is created successfully, or else return an error

Check if the mutex (which ensures thread safety) was created
successfully, and return an error if it wasn't. Moved this check to the
top of the function, so that it doesn't need to clean up memory in the
case of failure.

Since successfully creating the mutex is now a precondition for creating
the list, we can safely remove the NULL checks when using the mutex.