fix(log_router): ensure mutex is created successfully, or else return an error
Check if the mutex (which ensures thread safety) was created
successfully, and return an error if it wasn't. Moved this check to the
top of the <<>> function, so that it doesn't need to clean up allocated
memory.

Since successfully creating the mutex is now a precondition for creating
the list, we can safely remove the NULL checks when using the mutex.